Add project to DraftCard
