+++ src/components/cards/DraftCard.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { Draft } from "../../models/draft.model.ts";
 import dayjs from "dayjs";
+import ProjectTag from "../UI/ProjectTag.vue";
 
 defineProps<{
   draft: Draft;
@@ -9,12 +10,18 @@ defineProps<{
 
 <template>
   <div
-    class="cursor-grab rounded-lg bg-gray-400 px-4 py-2.5 active:cursor-grabbing"
+    class="cursor-grab space-y-1 rounded-lg bg-gray-400 px-4 py-2.5 active:cursor-grabbing"
   >
     <p class="">{{ draft.title }}</p>
-    <span class="text-xs text-gray-350">{{
-      dayjs(draft.dateCreated).format("DD MMM, HH:MM")
-    }}</span>
+    <div class="space-x flex items-center space-x-1">
+      <ProjectTag
+        v-if="draft.projectId"
+        :project-id="draft.projectId"
+      ></ProjectTag>
+      <span class="text-xs text-gray-350">
+        {{ dayjs(draft.dateCreated).format("DD MMM, HH:MM") }}
+      </span>
+    </div>
   </div>
 </template>
